# docs/swagger/swagger.yaml
basePath: /v1
definitions:
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      provider:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  http.ErrorResponse:
    properties:
      error:
        example: descriptive error message
        type: string
    type: object
  http.loginRequest:
    properties:
      email:
        example: joao@empresa.com
        type: string
      password:
        example: s3nh4segura
        type: string
    type: object
  http.registerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
  http.tokenResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API interna para o time de desenvolvimento.
  title: API Pão de Queijo
  version: "1.0"
paths:
  /auth/github:
    get:
      description: Redireciona para o GitHub OAuth
      responses:
        "302":
          description: Redireciona para GitHub
      summary: Iniciar login com GitHub
      tags:
      - auth
  /auth/github/callback:
    get:
      description: GitHub chama este endpoint após autorização
      parameters:
      - description: Código do GitHub
        in: query
        name: code
        required: true
        type: string
      - description: State anti-CSRF
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: token JWT
          schema:
            $ref: '#/definitions/http.tokenResponse'
        "400":
          description: State inválido
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Email já cadastrado com senha local
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "502":
          description: Erro no GitHub
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Callback do GitHub OAuth
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credenciais
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: token JWT
          schema:
            $ref: '#/definitions/http.tokenResponse'
        "401":
          description: Credenciais inválidas
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Login com email e senha
      tags:
      - auth
  /users:
    post:
      consumes:
      - application/json
      description: Cria um novo usuário com email e senha
      parameters:
      - description: Dados do usuário
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "409":
          description: Email já cadastrado
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Cadastrar usuário
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"